+++ src/traits.rs
@@ -25,6 +25,8 @@ use crate::{
 };
 #[cfg(feature = "arbitrary")]
 use arbitrary::Arbitrary;
+#[cfg(feature = "borsh")]
+use borsh::{BorshDeserialize, BorshSerialize};
 use bytemuck::{self, Pod, TransparentWrapper};
 use core::{
     fmt::{Binary, Debug, Display, LowerHex, Octal, UpperHex},
@@ -67,6 +69,7 @@ macro_rules! comment_features {
     ($comment:expr) => {
         #[cfg(all(
             not(feature = "arbitrary"),
+            not(feature = "borsh"),
             not(feature = "num-traits"),
             not(feature = "serde")
         ))]
@@ -77,6 +80,7 @@ macro_rules! comment_features {
 
         #[cfg(all(
             not(feature = "arbitrary"),
+            not(feature = "borsh"),
             not(feature = "num-traits"),
             feature = "serde"
         ))]
@@ -84,7 +88,7 @@ macro_rules! comment_features {
             $comment;
             pub trait FixedOptionalFeatures: Sealed
             where
-                Self: Serialize + for<'de> Deserialize<'de>
+                Self: Serialize + for<'de> Deserialize<'de>,
             {
             }
         }
@@ -92,6 +96,7 @@ macro_rules! comment_features {
         // Do *not* add MulAdd constaint, as it conflicts with Fixed::mul_add
         #[cfg(all(
             not(feature = "arbitrary"),
+            not(feature = "borsh"),
             feature = "num-traits",
             not(feature = "serde")
         ))]
@@ -112,7 +117,12 @@ macro_rules! comment_features {
         }
 
         // Do *not* add MulAdd constaint, as it conflicts with Fixed::mul_add
-        #[cfg(all(not(feature = "arbitrary"), feature = "num-traits", feature = "serde"))]
+        #[cfg(all(
+            not(feature = "arbitrary"),
+            not(feature = "borsh"),
+            feature = "num-traits",
+            feature = "serde"
+        ))]
         doc_comment! {
             $comment;
             pub trait FixedOptionalFeatures: Sealed
@@ -131,7 +141,87 @@ macro_rules! comment_features {
         }
 
         #[cfg(all(
+            not(feature = "arbitrary"),
+            feature = "borsh",
+            not(feature = "num-traits"),
+            not(feature = "serde")
+        ))]
+        doc_comment! {
+            $comment;
+            pub trait FixedOptionalFeatures: Sealed
+            where
+                Self: BorshSerialize + BorshDeserialize,
+            {}
+        }
+
+        #[cfg(all(
+            not(feature = "arbitrary"),
+            feature = "borsh",
+            not(feature = "num-traits"),
+            feature = "serde"
+        ))]
+        doc_comment! {
+            $comment;
+            pub trait FixedOptionalFeatures: Sealed
+            where
+                Self: BorshSerialize + BorshDeserialize,
+                Self: Serialize + for<'de> Deserialize<'de>,
+            {
+            }
+        }
+
+        // Do *not* add MulAdd constaint, as it conflicts with Fixed::mul_add
+        #[cfg(all(
+            not(feature = "arbitrary"),
+            feature = "borsh",
+            feature = "num-traits",
+            not(feature = "serde")
+        ))]
+        doc_comment! {
+            $comment;
+            pub trait FixedOptionalFeatures: Sealed
+            where
+                Self: BorshSerialize + BorshDeserialize,
+                Self: Zero + Bounded + Inv,
+                Self: CheckedAdd + CheckedSub + CheckedNeg + CheckedMul,
+                Self: CheckedDiv + CheckedRem + CheckedShl + CheckedShr,
+                Self: SaturatingAdd + SaturatingSub + SaturatingMul,
+                Self: WrappingAdd + WrappingSub + WrappingNeg + WrappingMul,
+                Self: WrappingShl + WrappingShr,
+                Self: OverflowingAdd + OverflowingSub + OverflowingMul,
+                Self: ToPrimitive + FromPrimitive + FloatConst,
+            {
+            }
+        }
+
+        // Do *not* add MulAdd constaint, as it conflicts with Fixed::mul_add
+        #[cfg(all(
+            not(feature = "arbitrary"),
+            feature = "borsh",
+            feature = "num-traits",
+            feature = "serde"
+        ))]
+        doc_comment! {
+            $comment;
+            pub trait FixedOptionalFeatures: Sealed
+            where
+                Self: BorshSerialize + BorshDeserialize,
+                Self: Zero + Bounded + Inv,
+                Self: CheckedAdd + CheckedSub + CheckedNeg + CheckedMul,
+                Self: CheckedDiv + CheckedRem + CheckedShl + CheckedShr,
+                Self: SaturatingAdd + SaturatingSub + SaturatingMul,
+                Self: WrappingAdd + WrappingSub + WrappingNeg + WrappingMul,
+                Self: WrappingShl + WrappingShr,
+                Self: OverflowingAdd + OverflowingSub + OverflowingMul,
+                Self: ToPrimitive + FromPrimitive + FloatConst,
+                Self: Serialize + for<'de> Deserialize<'de>,
+            {
+            }
+        }
+
+        #[cfg(all(
             feature = "arbitrary",
+            not(feature = "borsh"),
             not(feature = "num-traits"),
             not(feature = "serde")
         ))]
@@ -144,7 +234,12 @@ macro_rules! comment_features {
             }
         }
 
-        #[cfg(all(feature = "arbitrary", not(feature = "num-traits"), feature = "serde"))]
+        #[cfg(all(
+            feature = "arbitrary",
+            not(feature = "borsh"),
+            not(feature = "num-traits"),
+            feature = "serde"
+        ))]
         doc_comment! {
             $comment;
             pub trait FixedOptionalFeatures: Sealed
@@ -156,7 +251,12 @@ macro_rules! comment_features {
         }
 
         // Do *not* add MulAdd constaint, as it conflicts with Fixed::mul_add
-        #[cfg(all(feature = "arbitrary", feature = "num-traits", not(feature = "serde")))]
+        #[cfg(all(
+            feature = "arbitrary",
+            not(feature = "borsh"),
+            feature = "num-traits",
+            not(feature = "serde")
+        ))]
         doc_comment! {
             $comment;
             pub trait FixedOptionalFeatures: Sealed
@@ -175,12 +275,101 @@ macro_rules! comment_features {
         }
 
         // Do *not* add MulAdd constaint, as it conflicts with Fixed::mul_add
-        #[cfg(all(feature = "arbitrary", feature = "num-traits", feature = "serde"))]
+        #[cfg(all(
+            feature = "arbitrary",
+            not(feature = "borsh"),
+            feature = "num-traits",
+            feature = "serde"
+        ))]
         doc_comment! {
             $comment;
             pub trait FixedOptionalFeatures: Sealed
             where
                 Self: for<'a> Arbitrary<'a>,
+                Self: Zero + Bounded + Inv,
+                Self: CheckedAdd + CheckedSub + CheckedNeg + CheckedMul,
+                Self: CheckedDiv + CheckedRem + CheckedShl + CheckedShr,
+                Self: SaturatingAdd + SaturatingSub + SaturatingMul,
+                Self: WrappingAdd + WrappingSub + WrappingNeg + WrappingMul,
+                Self: WrappingShl + WrappingShr,
+                Self: OverflowingAdd + OverflowingSub + OverflowingMul,
+                Self: ToPrimitive + FromPrimitive + FloatConst,
+                Self: Serialize + for<'de> Deserialize<'de>,
+            {
+            }
+        }
+
+        #[cfg(all(
+            feature = "arbitrary",
+            feature = "borsh",
+            not(feature = "num-traits"),
+            not(feature = "serde")
+        ))]
+        doc_comment! {
+            $comment;
+            pub trait FixedOptionalFeatures: Sealed
+            where
+                Self: for<'a> Arbitrary<'a>,
+                Self: BorshSerialize + BorshDeserialize,
+            {
+            }
+        }
+
+        #[cfg(all(
+            feature = "arbitrary",
+            feature = "borsh",
+            not(feature = "num-traits"),
+            feature = "serde"
+        ))]
+        doc_comment! {
+            $comment;
+            pub trait FixedOptionalFeatures: Sealed
+            where
+                Self: for<'a> Arbitrary<'a>,
+                Self: BorshSerialize + BorshDeserialize,
+                Self: Serialize + for<'de> Deserialize<'de>
+            {
+            }
+        }
+
+        // Do *not* add MulAdd constaint, as it conflicts with Fixed::mul_add
+        #[cfg(all(
+            feature = "arbitrary",
+            feature = "borsh",
+            feature = "num-traits",
+            not(feature = "serde")
+        ))]
+        doc_comment! {
+            $comment;
+            pub trait FixedOptionalFeatures: Sealed
+            where
+                Self: for<'a> Arbitrary<'a>,
+                Self: BorshSerialize + BorshDeserialize,
+                Self: Zero + Bounded + Inv,
+                Self: CheckedAdd + CheckedSub + CheckedNeg + CheckedMul,
+                Self: CheckedDiv + CheckedRem + CheckedShl + CheckedShr,
+                Self: SaturatingAdd + SaturatingSub + SaturatingMul,
+                Self: WrappingAdd + WrappingSub + WrappingNeg + WrappingMul,
+                Self: WrappingShl + WrappingShr,
+                Self: OverflowingAdd + OverflowingSub + OverflowingMul,
+                Self: ToPrimitive + FromPrimitive + FloatConst,
+            {
+            }
+        }
+
+        // Do *not* add MulAdd constaint, as it conflicts with Fixed::mul_add
+        #[cfg(all(
+            feature = "arbitrary",
+            feature = "borsh",
+            feature = "num-traits",
+            feature = "serde"
+        ))]
+        doc_comment! {
+            $comment;
+            pub trait FixedOptionalFeatures: Sealed
+            where
+                Self: for<'a> Arbitrary<'a>,
+                Self: BorshSerialize + BorshDeserialize,
                 Self: Zero + Bounded + Inv,
                 Self: CheckedAdd + CheckedSub + CheckedNeg + CheckedMul,
                 Self: CheckedDiv + CheckedRem + CheckedShl + CheckedShr,
@@ -203,7 +392,10 @@ depending on the crate’s [optional features], and should not be used directly.
  1. If the `arbitrary` feature is enabled, [`Arbitrary`] is a supertrait of
     [`Fixed`].
 
- 2. If the `num-traits` experimental feature is enabled, the following
+ 2. If the `borsh` experimental feature is enabled, [`BorshSerialize`] and
+    [`BorshDeserialize`] are supertraits of [`Fixed`].
+
+ 3. If the `num-traits` experimental feature is enabled, the following
     are supertraits of [`Fixed`]:
 
       * [`Zero`]
@@ -234,7 +426,7 @@ depending on the crate’s [optional features], and should not be used directly.
     [`FixedSigned`] and [`FixedUnsigned`] because they have [`Num`] as
     a supertrait.
 
- 3. If the `serde` feature is enabled, [`Serialize`] and
+ 4. If the `serde` feature is enabled, [`Serialize`] and
     [`Deserialize`] are supertraits of [`Fixed`].
 
 [`MulAddAssign`]: num_traits::ops::mul_add::MulAddAssign